Split unit and acceptance tests into separate suites. Added npm run acceptance for running acceptance tests. Mocked HttpBackend in unit tests.
