+++ package.json
@@ -8,7 +8,8 @@
   },
   "scripts": {
     "build": "webpack",
-    "test": "karma start karma.conf.js"
+    "test": "karma start karma.conf.js ./src/**.spec.ts",
+    "acceptance": "karma start karma.conf.js ./src/**.spec.acceptance.ts"
   },
   "keywords": [
     "bugsplat",
